openai/code-explainer/requirements.txt
openai==0.27.6
python-dotenv=1.0.0
requests==2.30.0
streamlit==1.22.0